Move the people/:id endpoint to Action folder.
Notice that we were able to help the compiler out by telling it what the
type of the runSQL function is. That function is very polymorphic, and
we have to help the compiler out and tell it what its type is.

Also, notice that the ApiAction type is parameterized - i.e., it takes a
type variable, which allows us to specialize it.

Finally, notice that runSQL (fully saturated) is a value of the same
type that our entire "action" is. It has to be, we can't put something
there that is not of that same type, because what we are doing is
composing values of the same type to get something else of the same
type.